finished multiple type system

+++ Assets/Util/OneLiners.cs
@@ -6,7 +6,7 @@ namespace Core.Util
     {
         public static float Time => UnityEngine.Time.realtimeSinceStartup;
         public static byte FloatToByte(float x) => (byte)((Math.Clamp(x, -1, 1) + 1) / 2f * byte.MaxValue);
-        public static float ByteToFloat(byte x) => x / byte.MaxValue * 2 - 1;
+        public static float ByteToFloat(byte x) => ((float)x / byte.MaxValue * 2 - 1);
 
     }
 }
